requirements.txt
flask
flask-restx
pandas
numpy
pyyaml
pydantic